requirements.txt
alabaster==1.0.0
allure-pytest==2.13.5
allure-python-commons==2.13.5
anyio==4.6.0
attrs==24.2.0
babel==2.16.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.21.2
execnet==2.1.1
Faker==30.0.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
Jinja2==3.1.4
jsonpath-ng==1.6.1
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pipdeptree==2.23.4
platformdirs==4.3.6
playwright==1.47.0
pluggy==1.5.0
ply==3.11
pyee==12.0.0
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-base-url==2.1.0
pytest-playwright==0.5.2
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-autodoc-typehints==2.5.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
testit-adapter-pytest==3.2.3
testit-api-client==4.0.0
testit-python-commons==3.2.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
